requirements.txt
openai>=1.0.0
python-dotenv>=1.0.0
rich>=13.0.0
pygments>=2.15.0
tree-sitter>=0.20.0
tree-sitter-python>=0.20.0
tree-sitter-javascript>=0.20.0
tree-sitter-typescript>=0.20.0
gitpython>=3.1.0
chromadb>=0.4.0
watchdog>=3.0.0
numpy>=1.24.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
websockets>=12.0
python-multipart>=0.0.6
tiktoken>=0.5.0
anthropic>=0.18.0
psutil>=5.9.0